# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WireGuard Mesh Manager API",
    "description": "Backend API for managing WireGuard VPN networks and devices",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/backup/create": {
      "post": {
        "tags": [
          "backup",
          "backup"
        ],
        "summary": "Create Backup",
        "description": "Create a backup of all networks, locations, and devices.\n\nArgs:\n    password: Optional password for encryption. If empty and encrypt=True, a random password is generated.\n    description: Optional description for the backup\n    exported_by: Who is creating the backup\n    encrypt: Whether to encrypt the backup\n    db: Database session\n\nReturns:\n    BackupCreateResponse with backup data and optional password",
        "operationId": "create_backup_api_backup_create_post",
        "parameters": [
          {
            "name": "password",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "",
              "title": "Password"
            }
          },
          {
            "name": "description",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Description"
            }
          },
          {
            "name": "exported_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "api",
              "title": "Exported By"
            }
          },
          {
            "name": "encrypt",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Encrypt"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/backup/restore": {
      "post": {
        "tags": [
          "backup",
          "backup"
        ],
        "summary": "Restore Backup",
        "description": "Restore networks, locations, and devices from backup data.\n\nArgs:\n    request: Restore request containing backup data and options\n    db: Database session\n\nReturns:\n    BackupRestoreResponse with restore results",
        "operationId": "restore_backup_api_backup_restore_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackupRestoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupRestoreResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/backup/upload": {
      "post": {
        "tags": [
          "backup",
          "backup"
        ],
        "summary": "Upload Backup File",
        "description": "Upload and optionally restore from a backup file.\n\nArgs:\n    file: Backup file to upload\n    password: Password for encrypted backup\n    overwrite_existing: Whether to overwrite existing networks\n    dry_run: If True, only parse and validate without restoring\n    db: Database session\n\nReturns:\n    JSON response with backup info or restore results",
        "operationId": "upload_backup_file_api_backup_upload_post",
        "parameters": [
          {
            "name": "password",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Password"
            }
          },
          {
            "name": "overwrite_existing",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Overwrite Existing"
            }
          },
          {
            "name": "dry_run",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Dry Run"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_backup_file_api_backup_upload_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/backup/info": {
      "get": {
        "tags": [
          "backup",
          "backup"
        ],
        "summary": "Get Backup Info From Data",
        "description": "Get information about backup data without restoring it.\n\nArgs:\n    backup_data: Backup data to analyze\n    password: Password for encrypted backup\n\nReturns:\n    BackupInfoResponse with backup details",
        "operationId": "get_backup_info_from_data_api_backup_info_get",
        "parameters": [
          {
            "name": "password",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Password"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "title": "Backup Data"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupInfoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BackupCreateResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "exported_by": {
            "type": "string",
            "title": "Exported By"
          },
          "encrypted": {
            "type": "boolean",
            "title": "Encrypted"
          },
          "networks_count": {
            "type": "integer",
            "title": "Networks Count"
          },
          "locations_count": {
            "type": "integer",
            "title": "Locations Count"
          },
          "devices_count": {
            "type": "integer",
            "title": "Devices Count"
          },
          "password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          },
          "backup_data": {
            "type": "object",
            "title": "Backup Data"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at",
          "exported_by",
          "encrypted",
          "networks_count",
          "locations_count",
          "devices_count",
          "backup_data"
        ],
        "title": "BackupCreateResponse",
        "description": "Response model for backup creation."
      },
      "BackupInfoResponse": {
        "properties": {
          "encrypted": {
            "type": "boolean",
            "title": "Encrypted"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "exported_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exported At"
          },
          "exported_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exported By"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "networks_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Networks Count"
          },
          "locations_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locations Count"
          },
          "devices_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Devices Count"
          },
          "networks": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Networks",
            "default": []
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "encrypted",
          "version"
        ],
        "title": "BackupInfoResponse",
        "description": "Response model for backup information."
      },
      "BackupRestoreRequest": {
        "properties": {
          "backup_data": {
            "type": "object",
            "title": "Backup Data"
          },
          "password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          },
          "overwrite_existing": {
            "type": "boolean",
            "title": "Overwrite Existing",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "backup_data"
        ],
        "title": "BackupRestoreRequest",
        "description": "Request model for backup restoration."
      },
      "BackupRestoreResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "networks_created": {
            "type": "integer",
            "title": "Networks Created",
            "default": 0
          },
          "networks_updated": {
            "type": "integer",
            "title": "Networks Updated",
            "default": 0
          },
          "locations_created": {
            "type": "integer",
            "title": "Locations Created",
            "default": 0
          },
          "devices_created": {
            "type": "integer",
            "title": "Devices Created",
            "default": 0
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "BackupRestoreResponse",
        "description": "Response model for backup restoration."
      },
...
      "Body_upload_backup_file_api_backup_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_backup_file_api_backup_upload_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}